remove misleading section about python dns package
This section without clear reasoning is very confusing and doesn't make
much sense.

It's not true that python-dns needs to be installed on Tendrl Server
machine.

Package python-dns is required for tendrl-ansible to work, but this
could be pointed out in section related to tendrl-ansible. Moreover
tendrl ansible rpm package has a rpm dependency on python-dns, so one
doesn't even have to care about this manually.